+++ config.c
@@ -520,6 +520,9 @@ parse_section_cursor(const char *key, const char *value, struct config *conf,
     else if (strcmp(key, "hide-when-typing") == 0)
         conf->cursor.hide_when_typing = str_to_bool(value);
 
+    else if (strcmp(key, "indicate-when-selecting") == 0)
+        conf->cursor.indicate_when_selecting = str_to_bool(value);
+
     else if (strcmp(key, "color") == 0) {
         char *value_copy = strdup(value);
         const char *text = strtok(value_copy, " ");
@@ -1249,6 +1252,7 @@ config_load(struct config *conf, const char *conf_path, bool errors_are_fatal)
             .style = CURSOR_BLOCK,
             .blink = false,
             .hide_when_typing = false,
+            .indicate_when_selecting = true,
             .color = {
                 .text = 0,
                 .cursor = 0,