+++ terminal.c
@@ -2213,7 +2213,7 @@ term_xcursor_update_for_seat(struct terminal *term, struct seat *seat)
     const char *xcursor
         = seat->pointer.hidden ? XCURSOR_HIDDEN
         : term->is_searching ? XCURSOR_LEFT_PTR
-        : selection_enabled(term, seat) ? XCURSOR_TEXT
+        : (selection_enabled(term, seat) && term->conf->cursor.indicate_when_selecting) ? XCURSOR_TEXT
         : XCURSOR_LEFT_PTR;
 
     render_xcursor_set(seat, term, xcursor);