+++ CHANGELOG.md
@@ -25,7 +25,8 @@
 * **pipe-selected** key binding. Works like **pipe-visible** and
   **pipe-scrollback**, but only pipes the currently selected text, if
   any (https://codeberg.org/dnkl/foot/issues/51).
-* **cursor.hide-when-typing** option to `footrc.
+* **cursor.hide-when-typing** option to `footrc`.
+* **cursor.indicate-when-selecting** option to `footrc`.
 
 
 ### Deprecated